pathfinder/requirements.txt
# Core — used by the daily digest and GitHub Actions
python-jobspy>=1.1.0
pydantic>=2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
groq>=0.4.0
simple-salesforce>=1.12.0
